Moved error to utilities
