+++ src/HotChocolate/Core/src/Execution/ErrorHelper.cs
@@ -249,4 +249,22 @@ internal static class ErrorHelper
             .SetMessage("PersistedQueryNotFound")
             .SetCode(ErrorCodes.Execution.PersistedQueryNotFound)
             .Build();
+    
+    public static IError NoNullBubbling_ArgumentValue_NotAllowed(
+        ArgumentNode argument)
+    {
+        var errorBuilder = ErrorBuilder.New();
+
+        if (argument.Value.Location is not null)
+        {
+            errorBuilder.AddLocation(
+                argument.Value.Location.Line,
+                argument.Value.Location.Column);
+        }
+
+        errorBuilder.SetSyntaxNode(argument.Value);
+        errorBuilder.SetMessage(ErrorHelper_NoNullBubbling_ArgumentValue_NotAllowed);
+
+        return errorBuilder.Build();
+    }
 }
